unify curio and trio tests
