+++ pylintrc
@@ -66,7 +66,7 @@ score=no
 # all. Valid levels: HIGH, INFERENCE, INFERENCE_FAILURE, UNDEFINED
 confidence=
 enable=
-disable=missing-docstring,invalid-name,bad-continuation,unused-argument,import-error,too-many-locals,redefined-builtin,pointless-string-statement,too-many-branches,undefined-loop-variable,bare-except,protected-access,global-statement,no-self-use,too-many-arguments,too-few-public-methods,no-init,too-many-instance-attributes,no-else-return
+disable=missing-docstring,invalid-name,bad-continuation,unused-argument,import-error,too-many-locals,redefined-builtin,pointless-string-statement,too-many-branches,undefined-loop-variable,bare-except,protected-access,global-statement,no-self-use,too-many-arguments,too-few-public-methods,no-init,too-many-instance-attributes,too-many-public-methods,no-else-return
 
 
 [TYPECHECK]
